requirements.txt
faiss-cpu
sentence-transformers
pypdf 
langchain
langchain-transformers
streamlit
google-generativeai